Extract output config generation
