+++ components/telemetry-operator/internal/configbuilder/section_builder_test.go
@@ -1,0 +1,1 @@
+package configbuilder
