+++ components/telemetry-operator/internal/secret/secret_helper.go
@@ -4,8 +4,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"strings"
-
 	telemetryv1alpha1 "github.com/kyma-project/kyma/components/telemetry-operator/apis/telemetry/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -106,14 +104,4 @@ func FetchSecretData(referencedSecret corev1.Secret, valFrom telemetryv1alpha1.V
 		return data, nil
 	}
 	return data, fmt.Errorf("the key '%s' cannot be found in the given secret '%s'", valFrom.SecretKey.Key, referencedSecret.Name)
-}
-
-// GenerateVariableName generates env variable name for a given secret reference by concatenating pipeline name, namespace, secret name and secret key.
-// Dots and dashes are replaced by underscores to be compliant with env variable name requirements.
-func GenerateVariableName(secretRef telemetryv1alpha1.SecretKeyRef, pipelineName string) string {
-	result := fmt.Sprintf("%s_%s_%s_%s", pipelineName, secretRef.Namespace, secretRef.Name, secretRef.Key)
-	result = strings.ToUpper(result)
-	result = strings.Replace(result, ".", "_", -1)
-	result = strings.Replace(result, "-", "_", -1)
-	return result
 }
