+++ components/telemetry-operator/internal/webhook/dryrun/dryrun.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/kyma-project/kyma/components/telemetry-operator/internal/configbuilder"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -11,7 +12,6 @@ import (
 
 	"github.com/google/uuid"
 	telemetryv1alpha1 "github.com/kyma-project/kyma/components/telemetry-operator/apis/telemetry/v1alpha1"
-	"github.com/kyma-project/kyma/components/telemetry-operator/internal/fluentbit"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -24,7 +24,7 @@ type Config struct {
 	FluentBitBinPath       string
 	FluentBitPluginDir     string
 	FluentBitConfigMapName types.NamespacedName
-	PipelineConfig         fluentbit.PipelineConfig
+	PipelineConfig         configbuilder.PipelineConfig
 }
 
 type DryRunner struct {
