+++ components/telemetry-operator/internal/webhook/dryrun/testdata/expected/pipelines/dynamic-parsers/parsers.conf
@@ -5,4 +5,3 @@
     Time_Format %d/%b/%Y:%H:%M:%S %z
     Types       user:string pass:string
     Name parser-1
-
