+++ components/telemetry-operator/internal/webhook/dryrun/testdata/expected/pipelines/dynamic/logpipeline-1.conf
@@ -21,4 +21,3 @@
     storage.total_limit_size 1G
     tls on
     tls.verify on
-
