+++ components/telemetry-operator/internal/webhook/logpipeline/validation/output_validator.go
@@ -2,6 +2,7 @@ package validation
 
 import (
 	"fmt"
+	"github.com/kyma-project/kyma/components/telemetry-operator/internal/configbuilder"
 
 	telemetryv1alpha1 "github.com/kyma-project/kyma/components/telemetry-operator/apis/telemetry/v1alpha1"
 	"github.com/kyma-project/kyma/components/telemetry-operator/internal/fluentbit"
@@ -20,14 +21,13 @@ func NewOutputValidator() OutputValidator {
 }
 
 func (v *outputValidator) Validate(logPipeline *telemetryv1alpha1.LogPipeline) error {
-
 	section, err := fluentbit.ParseSection(logPipeline.Spec.Output.Custom)
 	if err != nil {
 		return err
 	}
 
-	if _, hasKey := section[fluentbit.OutputStorageMaxSizeKey]; hasKey {
-		return fmt.Errorf("log pipeline '%s' contains forbidden configuration key '%s'", logPipeline.Name, fluentbit.OutputStorageMaxSizeKey)
+	if _, hasKey := section[configbuilder.OutputStorageMaxSizeKey]; hasKey {
+		return fmt.Errorf("log pipeline '%s' contains forbidden configuration key '%s'", logPipeline.Name, configbuilder.OutputStorageMaxSizeKey)
 	}
 
 	return nil
